Move components into a crate
Also moved statistics.
