+++ crates/genesis_components/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+authors = ["Ben Askew-Gale <bengale2007@googlemail.com>"]
+name = "genesis_components"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+
+[dependencies]
+genesis_brain = { workspace = true }
+genesis_maths = { workspace = true }
+genesis_config = { workspace = true }
+genesis_color = { workspace = true }
+genesis_ecosystem = { workspace = true }
+bevy = { workspace = true }
+bevy_egui = { workspace = true }
+derive_more = { workspace = true }
+derive-getters = { workspace = true }
+itertools = { workspace = true }
+serde_derive = { workspace = true}
+serde = { workspace = true}
+
+[dev-dependencies]
+genesis_newtype = { workspace = true }
