+++ crates/genesis_components/src/lay.rs
@@ -1,0 +1,8 @@
+use bevy::prelude::Component;
+use derive_more::Deref;
+
+#[derive(Component)]
+pub struct TryingToLay;
+
+#[derive(Component, Copy, Clone, Debug, Deref, Ord, PartialEq, Eq, PartialOrd)]
+pub struct EggsLaid(pub usize);
