+++ crates/genesis_components/src/lib.rs
@@ -1,0 +1,133 @@
+use bevy::prelude::{Color, Component, Entity};
+use derive_more::{Add, Deref, DerefMut, From};
+use genesis_color::rgb_to_hex;
+use genesis_ecosystem::Energy;
+use genesis_maths::cantor_pairing;
+use serde_derive::Serialize;
+
+pub mod eat;
+pub mod grow;
+pub mod lay;
+pub mod mind;
+pub mod see;
+pub mod time;
+
+#[derive(Component, Debug, PartialEq, Eq, Deref, DerefMut, From, Add)]
+pub struct BurntEnergy(Energy);
+
+impl BurntEnergy {
+    pub const fn new() -> Self {
+        Self(Energy::new_empty())
+    }
+}
+
+impl BurntEnergy {
+    pub fn return_energy(&mut self) -> Energy {
+        let amount = self.amount();
+        self.take_energy(amount)
+    }
+}
+
+#[derive(Component, Debug)]
+pub struct MovementSum {
+    translation_sum: f32,
+    rotation_sum: f32,
+}
+
+impl MovementSum {
+    pub const fn new() -> Self {
+        Self {
+            translation_sum: 0.0,
+            rotation_sum: 0.0,
+        }
+    }
+
+    pub fn uint_portion(&mut self) -> usize {
+        let tran_floor = self.translation_sum.floor();
+        self.translation_sum -= tran_floor;
+
+        let rot_floor = self.rotation_sum.floor();
+        self.rotation_sum -= rot_floor;
+
+        (tran_floor + rot_floor) as usize
+    }
+
+    pub fn add_translation(&mut self, translation: f32, translation_cost: f32) {
+        self.translation_sum += translation.abs() * translation_cost;
+    }
+    pub fn add_rotation(&mut self, rotation: f32, rotation_cost: f32) {
+        self.rotation_sum += rotation.abs() * rotation_cost;
+    }
+}
+
+#[derive(Component, Debug)]
+pub struct ThinkingSum(f32);
+
+impl ThinkingSum {
+    pub const fn new() -> Self {
+        Self(0.0)
+    }
+
+    pub fn add_thought(&mut self, synapses: usize, cost: f32) {
+        self.0 += synapses as f32 * cost;
+    }
+
+    pub fn uint_portion(&mut self) -> usize {
+        let thought_floor = self.0.floor();
+        self.0 -= thought_floor;
+
+        thought_floor as usize
+    }
+}
+
+#[derive(Component, Debug)]
+pub struct Egg;
+
+#[derive(Component, Debug)]
+pub struct Hatching;
+
+#[derive(Component, Debug)]
+pub struct Juvenile;
+
+#[derive(Component, Debug)]
+pub struct Adult;
+
+#[derive(
+    Component, Debug, Deref, DerefMut, Clone, Copy, From, Add, Ord, PartialOrd, Eq, PartialEq,
+)]
+pub struct Generation(pub usize);
+
+#[derive(Debug, Component, Serialize, Clone)]
+pub struct Relations {
+    entity: (u32, String),
+    parent: Option<u32>,
+    children: Vec<u32>,
+}
+
+impl Relations {
+    pub fn new(entity: (Entity, Color), parent: Option<Entity>) -> Self {
+        let parent = parent.map(|e| cantor_pairing(e.generation(), e.index()));
+        Self {
+            entity: Self::convert(entity),
+            parent,
+            children: vec![],
+        }
+    }
+
+    pub fn add_child(&mut self, child: Entity) {
+        self.children
+            .push(cantor_pairing(child.generation(), child.index()))
+    }
+
+    pub fn is_interesting(&self) -> bool {
+        !(self.parent.is_none() && self.children.is_empty())
+    }
+
+    fn convert(input: (Entity, Color)) -> (u32, String) {
+        let (e, c) = input;
+        (
+            cantor_pairing(e.generation(), e.index()),
+            rgb_to_hex(c.r(), c.g(), c.b()),
+        )
+    }
+}
