+++ crates/genesis_components/src/time.rs
@@ -1,12 +1,10 @@
 use std::fmt;
 
 use bevy::{
-    prelude::{Component, Query, Res, ResMut, Resource},
-    time::{Stopwatch, Time},
+    prelude::{Component, Resource},
+    time::Stopwatch,
 };
 use derive_more::{Deref, DerefMut};
-
-use crate::{attributes, config, mind::MindOutput};
 
 #[derive(Resource, Debug, Deref, DerefMut)]
 pub struct SimulationTime(pub Stopwatch);
@@ -51,6 +49,12 @@ impl Heart {
     }
 }
 
+impl Default for Heart {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 #[derive(Component, Debug, Deref, DerefMut)]
 pub struct InternalTimer(pub Stopwatch);
 
@@ -69,36 +73,5 @@ impl Default for InternalTimer {
 impl fmt::Display for InternalTimer {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", self.elapsed().as_secs())
-    }
-}
-
-pub fn progress_age_system(time: Res<Time>, mut query: Query<&mut Age>) {
-    for mut age in query.iter_mut() {
-        age.tick(time.delta());
-    }
-}
-
-pub fn progress_timers_system(time: Res<Time>, mut query: Query<(&mut Heart, &mut InternalTimer)>) {
-    for (mut heart, mut internal_timer) in query.iter_mut() {
-        heart.tick(time.delta());
-        internal_timer.tick(time.delta());
-    }
-}
-
-pub fn progress_simulation_timer(time: Res<Time>, mut simulation_timer: ResMut<SimulationTime>) {
-    simulation_timer.tick(time.delta());
-}
-
-pub fn reset_internal_timer_system(
-    mut query: Query<(
-        &mut InternalTimer,
-        &MindOutput,
-        &attributes::InternalTimerBoundary,
-    )>,
-) {
-    for (mut internal_timer, mind_out, boundary) in query.iter_mut() {
-        if mind_out[config::RESET_TIMER_INDEX] > **boundary {
-            internal_timer.reset();
-        }
     }
 }
