+++ crates/genesis_lib/src/behaviour/growing.rs
@@ -1,14 +1,15 @@
 use bevy::{
-    prelude::{Commands, Component, Deref, DerefMut, Entity, Query, Res, With, Without},
+    prelude::{Commands, Entity, Query, Res, With, Without},
     sprite::Sprite,
     time::{Stopwatch, Time},
 };
 use bevy_rapier2d::prelude::Collider;
 
-use crate::{attributes, body, config, lifecycle::Egg, mind};
-
-#[derive(Component, Debug, Deref, DerefMut)]
-pub struct TryingToGrow(pub Stopwatch);
+use crate::{
+    attributes, body,
+    components::{grow::*, mind, Egg},
+    config,
+};
 
 type GrowerTest<'a> = (
     Entity,
@@ -33,26 +34,6 @@ pub fn process_growers_system(
         if mind_out[config::WANT_TO_GROWN_INDEX] <= **boundary {
             commands.entity(entity).remove::<TryingToGrow>();
         }
-    }
-}
-
-#[derive(Component, Debug)]
-pub struct GrowingSum(f32);
-
-impl GrowingSum {
-    pub const fn new() -> Self {
-        Self(0.0)
-    }
-
-    pub fn add_growing_time(&mut self, time: f32, rate: f32) {
-        self.0 += time * rate;
-    }
-
-    pub fn uint_portion(&mut self) -> usize {
-        let growing_floor = self.0.floor();
-        self.0 -= growing_floor;
-
-        growing_floor as usize
     }
 }
 
@@ -88,26 +69,6 @@ pub fn grow_bug_system(
         };
         sprite.custom_size = Some(vitality.size().sprite());
         *collider = vitality.size().collider();
-    }
-}
-
-#[derive(Component, Debug)]
-pub struct SizeSum(f32);
-
-impl SizeSum {
-    pub const fn new() -> Self {
-        Self(0.0)
-    }
-
-    pub fn add_existence_time(&mut self, time: f32, cost: f32) {
-        self.0 += time * cost;
-    }
-
-    pub fn uint_portion(&mut self) -> usize {
-        let size_floor = self.0.floor();
-        self.0 -= size_floor;
-
-        size_floor as usize
     }
 }
