+++ crates/genesis_lib/src/behaviour/mod.rs
@@ -3,16 +3,16 @@ use std::time::Duration;
 use bevy::prelude::{App, Plugin, SystemSet};
 use iyes_loopless::prelude::*;
 
-use crate::ui;
+use crate::{components::time, ui};
 
 pub mod eating;
-pub mod growth;
+pub mod growing;
 pub mod laying;
 pub mod metabolism;
-pub mod movement;
-pub mod sight;
+pub mod moving;
+pub mod seeing;
 pub mod thinking;
-pub mod timers;
+pub mod timing;
 
 pub fn before_thinking_system_set() -> SystemSet {
     ConditionSet::new()
@@ -20,7 +20,7 @@ pub fn before_thinking_system_set() -> SystemSet {
         .before("thinking")
         .run_if_not(ui::is_paused)
         .with_system(thinking::sensory_system)
-        .with_system(sight::process_sight_system)
+        .with_system(seeing::process_sight_system)
         .into()
 }
 
@@ -37,11 +37,11 @@ pub fn after_thinking_system_set() -> SystemSet {
         .label("after_thinking")
         .after("thinking")
         .run_if_not(ui::is_paused)
-        .with_system(timers::reset_internal_timer_system)
-        .with_system(movement::movement_system)
+        .with_system(timing::reset_internal_timer_system)
+        .with_system(moving::movement_system)
         .with_system(eating::process_eaters_system)
         .with_system(laying::process_layers_system)
-        .with_system(growth::process_growers_system)
+        .with_system(growing::process_growers_system)
         .into()
 }
 
@@ -51,7 +51,7 @@ pub fn attempting_behaviour_system_set() -> SystemSet {
         .after("after_thinking")
         .run_if_not(ui::is_paused)
         .with_system(eating::attempted_to_eat_system)
-        .with_system(growth::attempted_to_grow_system)
+        .with_system(growing::attempted_to_grow_system)
         .into()
 }
 
@@ -59,7 +59,7 @@ pub fn other_behaviour_system_set() -> SystemSet {
     ConditionSet::new()
         .label("other")
         .run_if_not(ui::is_paused)
-        .with_system(growth::existence_system)
+        .with_system(growing::existence_system)
         .with_system(eating::eating_system)
         .into()
 }
@@ -68,7 +68,7 @@ pub fn slow_behaviour_system_set() -> SystemSet {
     ConditionSet::new()
         .run_if_not(ui::is_paused)
         .with_system(laying::lay_egg_system)
-        .with_system(growth::grow_bug_system)
+        .with_system(growing::grow_bug_system)
         .into()
 }
 
@@ -83,9 +83,9 @@ pub fn very_slow_system_set() -> SystemSet {
 pub fn timers_system_set() -> SystemSet {
     ConditionSet::new()
         .run_if_not(ui::is_paused)
-        .with_system(timers::progress_simulation_timer)
-        .with_system(timers::progress_age_system)
-        .with_system(timers::progress_timers_system)
+        .with_system(timing::progress_simulation_timer)
+        .with_system(timing::progress_age_system)
+        .with_system(timing::progress_timers_system)
         .into()
 }
 
@@ -93,7 +93,7 @@ pub struct GenesisBehaviourPlugin;
 
 impl Plugin for GenesisBehaviourPlugin {
     fn build(&self, app: &mut App) {
-        app.insert_resource(timers::SimulationTime::default())
+        app.insert_resource(time::SimulationTime::default())
             .add_fixed_timestep(Duration::from_secs_f32(0.1), "slow")
             .add_fixed_timestep(Duration::from_secs_f32(1.0), "very_slow")
             .add_fixed_timestep_system_set("very_slow", 0, very_slow_system_set())
