+++ crates/genesis_lib/src/lifecycle.rs
@@ -1,30 +1,10 @@
-use bevy::prelude::{
-    Commands, Component, DespawnRecursiveExt, Entity, Query, ResMut, With, Without,
-};
-use derive_more::{Add, Deref, DerefMut, From};
+use bevy::prelude::{Commands, DespawnRecursiveExt, Entity, Query, ResMut, With, Without};
 
-use crate::{ancestors, attributes, behaviour::timers, body, ecosystem};
-
-#[derive(Component, Debug)]
-pub struct Egg;
-
-#[derive(Component, Debug)]
-pub struct Hatching;
-
-#[derive(Component, Debug)]
-pub struct Juvenile;
-
-#[derive(Component, Debug)]
-pub struct Adult;
-
-#[derive(
-    Component, Debug, Deref, DerefMut, Clone, Copy, From, Add, Ord, PartialOrd, Eq, PartialEq,
-)]
-pub struct Generation(pub usize);
+use crate::{attributes, body, components::*, ecosystem, statistics};
 
 pub fn transition_to_adult_system(
     mut commands: Commands,
-    bug_query: Query<(Entity, &timers::Age, &attributes::AdultAge), With<Juvenile>>,
+    bug_query: Query<(Entity, &time::Age, &attributes::AdultAge), With<Juvenile>>,
 ) {
     for (entity, age, adult_age) in bug_query.iter() {
         if age.elapsed_secs() > **adult_age {
@@ -35,7 +15,7 @@ pub fn transition_to_adult_system(
 
 pub fn transition_to_hatching_system(
     mut commands: Commands,
-    egg_query: Query<(Entity, &timers::Age, &attributes::HatchAge), (With<Egg>, Without<Hatching>)>,
+    egg_query: Query<(Entity, &time::Age, &attributes::HatchAge), (With<Egg>, Without<Hatching>)>,
 ) {
     for (entity, age, hatch_age) in egg_query.iter() {
         if age.elapsed_secs() > **hatch_age {
@@ -47,13 +27,13 @@ pub fn transition_to_hatching_system(
 pub fn kill_bug_system(
     mut commands: Commands,
     mut ecosystem: ResMut<ecosystem::Ecosystem>,
-    mut family_tree: ResMut<ancestors::FamilyTree>,
+    mut family_tree: ResMut<statistics::FamilyTree>,
     mut query: Query<(
         Entity,
         &mut body::Vitality,
         &attributes::DeathAge,
-        &timers::Age,
-        &ancestors::Relations,
+        &time::Age,
+        &Relations,
     )>,
 ) {
     for (entity, mut vitality, death_age, age, relations) in query.iter_mut() {
